change: moved Media Select as a standalone component
