+++ apps/web/src/app/app/components/initialization/media-initialization.tsx
@@ -1,9 +1,8 @@
 import VideoStream from "@/web/app/app/components/video-stream";
 import { Button } from "@/web/components/ui/button";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/web/components/ui/select";
-import { cn } from "@/web/lib/utils";
-import { Info, Mic, MicOff, SquarePlus, Video, VideoOff } from "lucide-react";
+import { Info, SquarePlus } from "lucide-react";
 import { useEffect, useState } from "react";
+import MediaSelect from "./components/media-select";
 
 interface MediaInitializationProps {
   onInitialize: (
@@ -13,57 +12,6 @@ interface MediaInitializationProps {
     videoDeviceId: string,
     audioDeviceId: string,
   ) => void;
-}
-
-interface MediaSelectProps {
-  mediaEnabled: boolean;
-  currentDeviceId: string | null;
-  inputDevices: MediaDeviceInfo[];
-  mediaToggleHandler: (value: boolean) => void;
-  onDeviceChange: (deviceId: string) => void;
-  kind: "video" | "audio";
-}
-
-const MediaSelect = ({
-  mediaEnabled,
-  mediaToggleHandler,
-  currentDeviceId,
-  inputDevices,
-  onDeviceChange,
-  kind,
-}: MediaSelectProps) => {
-  return <div className="flex flex-col gap-1 mb-2 mt-2">
-    <div className="flex gap-3 items-center">
-      <Button onClick={() => mediaToggleHandler(!mediaEnabled)} variant="outline" className={cn(mediaEnabled ? "bg-rose-400 text-white" : "")}>
-        {
-          mediaEnabled ? (
-            kind === "audio" ? <Mic /> : <Video />
-          ) : (
-            kind === "audio" ? <MicOff /> : <VideoOff />
-          )
-        }
-      </Button>
-
-      <Select 
-        value={currentDeviceId ?? ""} 
-        disabled={currentDeviceId === null}
-        onValueChange={onDeviceChange}
-      >
-        <SelectTrigger className="w-[20rem]">
-          <SelectValue placeholder="Select device" />
-        </SelectTrigger>
-        <SelectContent>
-          {
-            inputDevices.map(device => (
-              <SelectItem key={device.deviceId} value={device.deviceId} className="capitalize">
-                {device.label}
-              </SelectItem>
-            ))
-          }
-        </SelectContent>
-      </Select>
-    </div>
-  </div> 
 }
 
 export default function MediaInitialization({
